% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_fleck.R
\name{geom_fleck}
\alias{geom_fleck}
\title{Fleck}
\usage{
geom_fleck(mapping = NULL, data = NULL, stat = "aspect_shading",
  position = "identity", res = 1, ..., na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE)
}
\arguments{
\item{...}{other arguments passed on to [layer()]. These are
often aesthetics, used to set an aesthetic to a fixed value, like
`color = "red"` or `size = 3`. They may also be parameters
to the paired geom/stat.}

\item{na.rm}{If `FALSE`, the default, missing values are removed with
a warning. If `TRUE`, missing values are silently removed.}
}
\description{
Fleck
}
