% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_slope_shading.R
\name{stat_slope_shading}
\alias{stat_slope_shading}
\title{Compute slope shading per group for the given data}
\usage{
stat_slope_shading(mapping = NULL, data = NULL, geom = "point",
  position = "identity", sun.elevation = 1/3 * pi, res = 1, ...,
  show.legend = NA, inherit.aes = TRUE)
}
\arguments{
\item{...}{other arguments passed on to [layer()]. These are
often aesthetics, used to set an aesthetic to a fixed value, like
`color = "red"` or `size = 3`. They may also be parameters
to the paired geom/stat.}
}
\description{
The aspect shading statistic transform the height data into aspect shading data value.
}
\examples{
# Example with the volcano dataset


}
