% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_sphere_shading.R
\name{stat_sphere_shading}
\alias{stat_sphere_shading}
\title{Compute sphere shading per group for the given data}
\usage{
stat_sphere_shading(mapping = NULL, data = NULL, geom = "point",
  position = "identity", sun.angle = 1/3 * pi, slope = F, res = 1,
  zscale = 1, ..., show.legend = NA, inherit.aes = TRUE)
}
\arguments{
\item{...}{other arguments passed on to [layer()]. These are
often aesthetics, used to set an aesthetic to a fixed value, like
`color = "red"` or `size = 3`. They may also be parameters
to the paired geom/stat.}
}
\description{
The sphere shading statistic transform the height data into sphere shading data value.
See ?ray_shade from the rayshader package.
}
\examples{
# Example with the volcano dataset


}
