% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_fleck.R
\name{geom_fleck}
\alias{geom_fleck}
\title{Fleck}
\usage{
geom_fleck(mapping = NULL, data = NULL, stat = "aspect_shading",
  position = "identity", res = 1, ..., na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE)
}
\arguments{
\item{...}{other arguments passed on to [layer()]. These are
often aesthetics, used to set an aesthetic to a fixed value, like
`color = "red"` or `size = 3`. They may also be parameters
to the paired geom/stat.}

\item{na.rm}{If `FALSE`, the default, missing values are removed with
a warning. If `TRUE`, missing values are silently removed.}
}
\description{
Fleck is an extension of ggplot designed to plot geographic data. This is inspired by the designs of James Cheshire and Oliver Uberti.
The geom_fleck draws multiple points randomly distributed on a grid defined by the aes x and y with a maximum density defined by the aesthetic density.
It can be used with stat = aspect_shading (option by default) or stat = identity that do not compute the shading value from the height z.
}
\examples{

ggplot(volcano2, aes(x, y, z = z)) + geom_fleck()+ coord_fixed() + theme_void()


}
